refactor all JS  component to JSX
