+++ front-end/src/components/Navbar.jsx
@@ -1,7 +1,7 @@
 /* Simple Navbar that will sit at the top of the page with the title/logo sitting to left and 'Login' Or 'Sign Up' button sitting to the right
  */
 import "./NavBar.css";
-import Button from "./Button.js";
+import Button from "./Button.jsx";
 
 function Navbar() {
   const func = () => {
